Changed display update
Lines are updated individually now and not the complete display in a fixed interval (Lines are now uasyncio tasks not only the method which updated the whole display)
Scrolling speed is now a range which can be cycled through with Button X as before
Scrolling of a line is now randomized in the current range each time the line is extended with a start part to now stay the same speed all the time

Color change is now no more instant but effective after the current delay for the scrolling speed is reached. Looks a bit strange but perhaps only for me or not used to it yet. Let's see if this stays. Would have to find a way to interrupt the running asyncio sleep.