Added instant color change and scrolling speed
When cycling through colors they are now displayed instantly. Before that the next delay for the scrolling speed was awaited and then changed. This led to a strange looking delayed color change especially when scrolling speed was slow. Now the current matrix state is cached and when a color change is triggered, the display is updated instantly and independent of the delay.

When cycling through scrolling speed the are now changed instantly. Before the scrolling speed was updated only when it was randomized in the given delay range. That happened only when a starting part was added to the current line. Now, the scrolling speed of each line is updated instantly and independent of the remaining time until the next planned randomization.

Moved the constants for delays, colors and compositions out of the Matrix class.